TRINITY: enable two PPT layers
https://jira.ccpgames.com/browse/GFX-1411

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1067975]

+++ Eve/SpaceObjectFactory/EveSOFDNA.h
@@ -97,7 +97,7 @@ public:
 	const EveSOFDataMgr::FactionSpriteSetColorData* GetFactionSpriteSetData( int groupIndex ) const;
 
 	// get pattern data
-	const EveSOFDataMgr::PatternProjectionData* GetPatternProjectionData( const char* hullName, uint32_t layer ) const;
+	const EveSOFDataMgr::PatternProjectionData* GetPatternProjectionData( const char* hullName, size_t layer ) const;
 	const EveSOFDataMgr::PatternData* GetPatternData() const;
 
 	// get mixed data
